fix bags at mappers

+++ src/main/java/org/example/controller/TaskController.java
@@ -32,7 +32,7 @@ public class TaskController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<TaskResponse> getTaskById(@PathVariable Long id) {
+    public ResponseEntity<TaskResponse> getTaskById(@PathVariable Integer id) {
         return ResponseEntity.ok(taskService.getTask(id));
     }
 }
