+++ src/main/java/org/example/repository/TestCaseRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TestCaseRepository extends JpaRepository<TestCase, Long> {
+public interface TestCaseRepository extends JpaRepository<TestCase, Integer> {
 }
